job restart from checkpoint bug fixed #2438
