+++ streampark-common/src/main/java/org/apache/streampark/common/enums/ApplicationType.java
@@ -19,30 +19,38 @@ package org.apache.streampark.common.enums;
 
 import java.io.Serializable;
 
-public enum Semantic implements Serializable {
+public enum ApplicationType implements Serializable {
+  /** StreamPark Flink */
+  STREAMPARK_FLINK(1, "StreamPark Flink"),
+  /** Apache Flink */
+  APACHE_FLINK(2, "Apache Flink"),
+  /** StreamPark Spark */
+  STREAMPARK_SPARK(3, "StreamPark Spark"),
+  /** Apache Spark */
+  APACHE_SPARK(4, "Apache Spark");
 
-    /**
-     *
-     */
-    EXACTLY_ONCE,
+  private final int type;
+  private final String name;
 
-    /**
-     *
-     */
-    AT_LEAST_ONCE,
+  ApplicationType(int type, String name) {
+    this.type = type;
+    this.name = name;
+  }
 
-    /**
-     *
-     */
-    NONE;
+  public int getType() {
+    return type;
+  }
 
-    public static Semantic of(String name) {
-        for (Semantic semantic : Semantic.values()) {
-            if (name.equals(semantic.name())) {
-                return semantic;
-            }
-        }
-        return null;
+  public String getName() {
+    return name;
+  }
+
+  public static ApplicationType of(int type) {
+    for (ApplicationType appType : ApplicationType.values()) {
+      if (appType.getType() == type) {
+        return appType;
+      }
     }
-
+    return null;
+  }
 }
